backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
alembic==1.12.0
python-multipart==0.0.6
